requirements.txt
streamlit
openai-whisper
numpy
torchaudio